Move validation code from reducer into action creator and get rid reduce_validate_entry() by doing that
