+++ core_api/src/lib.rs
@@ -419,7 +419,7 @@ mod tests {
         // Check in holochain instance's history that the commit event has been processed
         // @TODO don't use history length in tests
         // @see https://github.com/holochain/holochain-rust/issues/195
-        assert_eq!(hc.state().unwrap().history.len(), 7);
+        assert_eq!(hc.state().unwrap().history.len(), 6);
     }
 
     #[test]
